backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.36
alembic==1.14.0
redis==5.2.0
celery==5.4.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
python-decouple==3.8
email-validator==2.2.0
jinja2==3.1.4
aiofiles==24.1.0
python-dateutil==2.8.2
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
isort==5.13.2

# Content Engine dependencies
aiohttp==3.12.15
requests==2.32.3
beautifulsoup4==4.12.3
pytrends==4.9.2
flake8==7.1.1
